refactor: authDropdown 별도 컴포넌트로 추출
